+++ Dockerfile
@@ -17,12 +17,12 @@ WORKDIR /code
 ENV GRADLE_USER_HOME=/code/.gradlecache \
    GRADLE_OPTS="-Djdk.lang.Process.launchMechanism=vfork -Dorg.gradle.vfs.watch=false"
 
-COPY build.gradle.kts settings.gradle.kts gradle.properties /code/
+COPY radar-gateway/build.gradle.kts settings.gradle.kts gradle.properties /code/
 COPY buildSrc /code/buildSrc
 
 RUN gradle downloadDependencies copyDependencies startScripts
 
-COPY src/ /code/src
+COPY radar-gateway/src/ /code/src
 
 RUN gradle jar
