+++ radar-gateway/src/integrationTest/docker/etc/config/oauth_client_details.csv
