+++ radar-gateway/src/main/kotlin/org/radarbase/gateway/inject/ProcessAvro.kt
