+++ radar-gateway/src/main/kotlin/org/radarbase/gateway/io/BinaryToAvroConverter.kt
