+++ radar-gateway/src/main/kotlin/org/radarbase/gateway/io/LimitedInputStream.kt
