+++ radar-gateway/src/main/kotlin/org/radarbase/gateway/kafka/KafkaHealthMetric.kt
