+++ radar-gateway/src/test/kotlin/org/radarbase/gateway/io/AvroRecordProcessorTest.kt
